requirements.txt
ArchivesSnake==0.9.1
boto3==1.17.35
botocore==1.20.35
bottle==0.12.19
gevent==21.1.2
jpylyzer==2.0.0
lxml==4.6.3
plac==1.3.3
python-decouple==3.4
requests==2.25.1
sh==1.14.1
